Rip types out of machinery
